# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MTranCore Translation Service API",
    "description": "Translation service supporting HTTP/REST, gRPC, and WebSocket protocols for machine translation using Bergamot engine",
    "version": "1.0.0",
    "contact": {
      "name": "MTranCore",
      "url": "https://github.com/xxnuo/MTranCore"
    },
    "license": {
      "name": "License",
      "url": "https://github.com/xxnuo/MTranCore/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8988",
      "description": "Unified Server - HTTP/REST, gRPC, and WebSocket (all on port 8988)"
    }
  ],
  "tags": [
    {
      "name": "HTTP",
      "description": "HTTP/REST API endpoints"
    },
    {
      "name": "gRPC",
      "description": "gRPC service methods"
    },
    {
      "name": "WebSocket",
      "description": "WebSocket protocol messages"
    },
    {
      "name": "Health",
      "description": "Health check endpoints"
    },
    {
      "name": "Lifecycle",
      "description": "Engine lifecycle management (poweron/exit)"
    },
    {
      "name": "Translation",
      "description": "Translation operations"
    }
  ],
  "paths": {


    "/exit": {
      "post": {
        "tags": ["HTTP", "Lifecycle"],
        "summary": "Power off server (HTTP)",
        "description": "Gracefully shut down the server with optional delay and force flag",
        "operationId": "exitHttp",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExitRequest"
              },
              "examples": {
                "immediate": {
                  "summary": "Immediate shutdown",
                  "value": {
                    "time": 0,
                    "force": false
                  }
                },
                "delayed": {
                  "summary": "Delayed shutdown (5 seconds)",
                  "value": {
                    "time": 5,
                    "force": false
                  }
                },
                "force": {
                  "summary": "Force shutdown",
                  "value": {
                    "time": 0,
                    "force": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shutdown initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["HTTP", "Lifecycle"],
        "summary": "Check engine health status (HTTP)",
        "description": "Check if the translation engine is loaded and health to process requests",
        "operationId": "healthHttp",
        "responses": {
          "200": {
            "description": "Health status",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/StandardResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/HealthResponse"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "health": {
                    "summary": "Engine is health",
                    "value": {
                      "code": 200,
                      "data": {
                        "health": true
                      }
                    }
                  },
                  "not_health": {
                    "summary": "Engine is not health",
                    "value": {
                      "code": 200,
                      "data": {
                        "health": false
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trans": {
      "post": {
        "tags": ["HTTP", "Translation"],
        "summary": "Translate text (HTTP)",
        "description": "Translate text using the loaded translation engine",
        "operationId": "transHttp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransRequest"
              },
              "examples": {
                "plain_text": {
                  "summary": "Plain text translation",
                  "value": {
                    "text": "Hello, world!",
                    "html": false
                  }
                },
                "html_text": {
                  "summary": "HTML text translation",
                  "value": {
                    "text": "<p>Hello, <strong>world</strong>!</p>",
                    "html": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Translation result (plain text)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "success": {
                    "summary": "Successful translation",
                    "value": "你好，世界！"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "500": {
            "description": "Translation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "503": {
            "description": "Server is shutting down",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/grpc/TranslatorService/Health": {
      "post": {
        "tags": ["gRPC", "Health"],
        "summary": "Health check (gRPC)",
        "description": "gRPC health check method",
        "operationId": "healthGrpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Health response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
        "/grpc/TranslatorService/Exit": {
      "post": {
        "tags": ["gRPC", "Lifecycle"],
        "summary": "Power off server (gRPC)",
        "description": "Shut down the gRPC server",
        "operationId": "exitGrpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exit response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/grpc/TranslatorService/Health": {
      "post": {
        "tags": ["gRPC", "Lifecycle"],
        "summary": "Check engine health status (gRPC)",
        "description": "Get the current engine status",
        "operationId": "healthGrpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Health response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseGrpc"
                }
              }
            }
          }
        }
      }
    },
    "/grpc/TranslatorService/Trans": {
      "post": {
        "tags": ["gRPC", "Translation"],
        "summary": "Translate text (gRPC)",
        "description": "Translate a single text using gRPC",
        "operationId": "transGrpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransRequestGrpc"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trans response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransResponseGrpc"
                }
              }
            }
          }
        }
      }
    },
    "/grpc/TranslatorService/TransStream": {
      "post": {
        "tags": ["gRPC", "Translation"],
        "summary": "Translate text stream (gRPC)",
        "description": "Translate multiple texts in streaming fashion using bidirectional gRPC stream",
        "operationId": "transStreamGrpc",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Stream of TransRequest messages",
                "properties": {
                  "stream": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/TransRequestGrpc"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream of trans responses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stream": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TransResponseGrpc"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ws": {
      "get": {
        "tags": ["WebSocket"],
        "summary": "WebSocket endpoint",
        "description": "WebSocket connection endpoint for bidirectional communication. Supports poweron, exit, health, and trans message types.",
        "operationId": "websocketConnect",
        "parameters": [
          {
            "name": "Connection",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Upgrade"]
            }
          },
          {
            "name": "Upgrade",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["websocket"]
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols - WebSocket connection established",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/WSResponse"
                    }
                  ]
                },
                "examples": {
                  "poweron_success": {
                    "summary": "Poweron success response",
                    "value": {
                      "type": "poweron",
                      "code": 200,
                      "msg": "success",
                      "data": {
                        "message": "Engine loaded successfully"
                      }
                    }
                  },
                  "trans_success": {
                    "summary": "Trans success response",
                    "value": {
                      "type": "trans",
                      "code": 200,
                      "msg": "success",
                      "data": {
                        "translated_text": "你好，世界！"
                      }
                    }
                  },
                  "health_response": {
                    "summary": "Health status response",
                    "value": {
                      "type": "health",
                      "code": 200,
                      "msg": "success",
                      "data": {
                        "health": true
                      }
                    }
                  }
                }
              }
            }
          },
          "426": {
            "description": "Upgrade Required - Not a WebSocket request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {

      "ExitRequest": {
        "type": "object",
        "properties": {
          "time": {
            "type": "integer",
            "format": "int32",
            "description": "Seconds to wait before shutdown",
            "default": 0,
            "minimum": 0,
            "example": 5
          },
          "force": {
            "type": "boolean",
            "description": "Force shutdown without waiting for requests",
            "default": false,
            "example": false
          }
        }
      },
      "TransRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "description": "Text to translate",
            "example": "Hello, world!"
          },
          "html": {
            "type": "boolean",
            "description": "Whether the text contains HTML markup",
            "default": false,
            "example": false
          }
        }
      },
      "TransRequestGrpc": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "description": "Text to translate"
          },
          "html": {
            "type": "boolean",
            "description": "Whether the text contains HTML markup",
            "default": false
          }
        }
      },
      "TransResponse": {
        "type": "object",
        "properties": {
          "translated_text": {
            "type": "string",
            "description": "Translated text",
            "example": "你好，世界！"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "health": {
            "type": "boolean",
            "description": "Whether the engine is health",
            "example": true
          }
        }
      },
      "StandardResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "Response code (0 for success, non-zero for errors)",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "Response message (usually present for errors)",
            "example": "Success"
          },
          "data": {
            "type": "object",
            "description": "Response data (present for successful requests)",
            "additionalProperties": true
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "OK"
          }
        }
      },

      "ExitResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "Server is shutting down"
          }
        }
      },
      "HealthResponseGrpc": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "OK"
          },
          "health": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "TransResponseGrpc": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "OK"
          },
          "translated_text": {
            "type": "string",
            "example": "你好，世界！"
          }
        }
      },
      "WSMessage": {
        "type": "object",
        "description": "WebSocket request message",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["poweron", "exit", "health", "trans"],
            "description": "Message type",
            "example": "trans"
          },
          "data": {
            "type": "object",
            "description": "Message data, structure depends on type",
            "oneOf": [
              {
                "$ref": "#/components/schemas/PoweronRequest"
              },
              {
                "$ref": "#/components/schemas/ExitRequest"
              },
              {
                "$ref": "#/components/schemas/TransRequest"
              },
              {
                "type": "object",
                "description": "Empty object for health message"
              }
            ]
          }
        },
        "examples": [
          {
            "type": "poweron",
            "data": {
              "path": "models/enzh"
            }
          },
          {
            "type": "trans",
            "data": {
              "text": "Hello, world!",
              "html": false
            }
          },
          {
            "type": "health",
            "data": {}
          }
        ]
      },
      "WSResponse": {
        "type": "object",
        "description": "WebSocket response message",
        "required": ["type", "code", "msg"],
        "properties": {
          "type": {
            "type": "string",
            "description": "Same as request type",
            "example": "trans"
          },
          "code": {
            "type": "integer",
            "description": "Response code (0 for success)",
            "example": 0
          },
          "msg": {
            "type": "string",
            "description": "Response message",
            "example": "success"
          },
          "data": {
            "type": "object",
            "description": "Response data (optional)",
            "additionalProperties": true
          }
        }
      },
      "ErrorCodes": {
        "type": "object",
        "description": "Standard error codes used across all protocols",
        "properties": {
          "0": {
            "type": "string",
            "description": "Success",
            "example": "Success"
          },
          "1000": {
            "type": "string",
            "description": "Poweron: Invalid parameters"
          },
          "1001": {
            "type": "string",
            "description": "Poweron: Path does not exist"
          },
          "1002": {
            "type": "string",
            "description": "Poweron: Incomplete model files"
          },
          "1003": {
            "type": "string",
            "description": "Poweron: Internal error"
          },
          "1009": {
            "type": "string",
            "description": "Poweron: Unknown error"
          },
          "1100": {
            "type": "string",
            "description": "Exit: Invalid parameters"
          },
          "1101": {
            "type": "string",
            "description": "Exit: Waiting for tasks to complete"
          },
          "1109": {
            "type": "string",
            "description": "Exit: Internal error"
          },
          "1200": {
            "type": "string",
            "description": "Trans: Invalid parameters"
          },
          "1201": {
            "type": "string",
            "description": "Trans: Translation failure"
          },
          "1209": {
            "type": "string",
            "description": "Trans: Internal error"
          }
        }
      }
    }
  }
}